packages/vadocs/pyproject.toml
[project]
name = "vadocs"
version = "0.1.0"
description = "Documentation validation engine with YAML frontmatter sync"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Vadim Rudakov"}]
keywords = ["documentation", "validation", "adr", "frontmatter", "yaml", "rag"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

# NO entry points - direct imports only in v0.1.0
# NO CLI script - library only in v0.1.0

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vadocs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
